+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGE HISTORY
 LATEST
 ------
 - Added support for ColouredText titles in MultiColumnLIstBox.
+- Added gutter option to Layout.
 - Fixed bug where moving focus between Frames resulted in no current focus.
 - Fixed internal state of RadioButton values to be consistent with selection.